Deployment/requirements.txt
Flask
Flask_Cors
python-dotenv
Requests
